# docs/openapi.yaml
openapi: 3.0.3
info:
  title: gloom-sistant
  version: 0.0.1
servers:
  - url: http://localhost:8000
    description: Local Server
paths:
...
  /players:
    get:
      tags:
        - Players
      summary: Returns a list of players.
      description: >
        Returns a list of basic info for all players.
      operationId: getPlayers
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    firstName:
                      type: string
                    lastName:
                      type: string
                    email:
                      type: string
              examples:
                normal:
                  summary: Normal
                  value: >
                    [
                      {
                        "id": 1,
                        "firstName": "Logan",
                        "lastName": "Lucky",
                        "email": "ll@llmail.com"
                      },
                      {
                        "id": 2,
                        "firstName": "George",
                        "lastName": "Gorgeous",
                        "email": "gg@ggmail.com"
                      }
                    ]
                empty:
                  summary: Empty
                  value: []
        404:
          $ref: '#/components/responses/NotFoundError'
        default:
          $ref: '#/components/responses/UnexpectedError'
    post:
      tags:
        - Players
      summary: Creates a player/user
      description: >
        Creates a player/user given required information and returns the id
        of the newly created record.
      operationId: createPlayer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
              required:
                - firstName
                - lastName
                - email
            examples:
              normal:
                summary: Normal
                value: >
                  {
                    "firstName": "Logan",
                    "lastName": "Lucky",
                    "email": "ll@llmail.com"
                  }
      responses:
        201:
          description: Resource created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
              examples:
                normal:
                  summary: Normal
                  value: >
                    {
                      "id": 1
                    }
        400:
          $ref: '#/components/responses/BadRequestError'
        default:
          $ref: '#/components/responses/UnexpectedError'
  /players/{id}:
    get:
      tags:
        - Players
      summary: Responds with a specific player's details
      description: >
        Returns all the details associated with a given player
      operationId: getPlayerDetail
      parameters:
        - name: id
          in: path
          description: ID of requested player
          required: true
          schema:
            type: integer
          examples:
            normal:
              summary: Normal
              value: 1
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  firstName:
                    type: string
                  lastName:
                    type: string
                  email:
                    type: string
                  characters:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        name:
                          type: string
              examples:
                normal:
                  summary: Normal
                  value: >
                    {
                      "id": 2,
                      "firstName": "George",
                      "lastName": "Gorgeous",
                      "email": "gg@ggmail.com",
                      "characters": [
                        {
                          "id": 3,
                          "name": "Turkish"
                        },
                        {
                          "id": 4,
                          "name": "Bullet Tooth Tony"
                        },
                        {
                          "id": 5,
                          "name": "Mickey O'Neil"
                        }
                      ]
                    }
        404:
          $ref: '#/components/responses/NotFoundError'
        default:
          $ref: '#/components/responses/UnexpectedError'
    put:
      tags:
        - Players
      summary: Updates a specific player's details
      description: >
        Updates the details of a specific player given new info in
        request payload.
      operationId: updatePlayer
      parameters:
        - name: id
          in: path
          description: ID of requested player
          required: true
          schema:
            type: integer
          examples:
            normal:
              summary: Normal
              value: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
      responses:
        204:
          description: Resource updated
        404:
          description: An error occurred
    delete:
      tags:
        - Players
      summary: Deletes a player
      description: >
        Deletes a player
      operationId: deletePlayer
      parameters:
        - name: id
          in: path
          description: ID of player to be deleted
          required: true
          schema:
            type: integer
          examples:
            normal:
              summary: Normal
              value: 1
      responses:
        204:
          description: Resource deleted
        404:
          $ref: '#/components/responses/NotFoundError'
        default:
          $ref: '#/components/responses/UnexpectedError'
...
  /campaigns:
    get:
      tags:
        - Campaigns
      summary: Returns a list of campaigns.
      description: >
        Returns a list of basic info for all campaigns.
      operationId: getCampaigns
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    name:
                      type: string
              examples:
                normal:
                  summary: Normal
                  value: >
                    [
                      {
                        "id": 1,
                        "name": "Jeres Game",
                        "characters": [
                          "1",
                          "3"
                        ]
                      },
                      {
                        "id": 2,
                        "name": "Gloomy Game",
                        "characters": [
                          "2",
                          "3"
                        ]
                      }
                    ]
                empty:
                  summary: Empty
                  value: []
        404:
          $ref: '#/components/responses/NotFoundError'
        default:
          $ref: '#/components/responses/UnexpectedError'
    
    post:
      tags:
        - Campaigns
      summary: Creates a campaign
      description: >
        Creates a campaign given required information and returns the id
        of the newly created record.
      operationId: createCampaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                characters:
                  type: array
                  items:
                    type: integer
              required:
                - name
            examples:
              normal:
                summary: Normal
                value: >
                  {
                    "name": "Logan",
                    "characters": [
                      1,
                      2
                    ]
                  }
      responses:
        201:
          description: Resource created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
              examples:
                normal:
                  summary: Normal
                  value: >
                    {
                      "id": 1
                    }
        400:
          $ref: '#/components/responses/BadRequestError'
        default:
          $ref: '#/components/responses/UnexpectedError'
  /campaigns/{id}:
    get:
      tags:
        - Campaigns
      summary: Responds with a specific campaigns's details
      description: >
        Returns all the details associated with a given campaign
      operationId: getCampaignDetail
      parameters:
        - name: id
          in: path
          description: ID of requested campaign
          required: true
          schema:
            type: integer
          examples:
            normal:
              summary: Normal
              value: 1
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  missionsCompleted:
                    type: array
                    items:
                      type: integer
                  missionsAvailable:
                    type: array
                    items:
                      type: integer
                  eventsCompleted:
                    type: array
                    items:
                      type: integer
                  itemsAvailable:
                    type: array
                    items:
                      type: integer
              examples:
                normal:
                  summary: Normal
                  value: >
                    [
                      {
                        "id": 1,
                        "name": "Jeres Game",
                        "gameMasterId": 1,
                        "missionsCompleted": [
                          "1",
                          "2",
                          "3",
                          "4"
                        ],
                        "missionsAvailable": [
                          "5"
                        ],
                        "eventsCompleted": [
                          "8"
                        ],
                        "itemsAvailable": [
                          "1",
                          "2",
                          "2",
                          "3",
                          "3",
                          "4",
                          "4",
                          "5",
                          "5",
                          "6",
                          "6",
                          "7",
                          "7",
                          "8",
                          "8",
                          "9",
                          "9",
                          "10",
                          "10",
                          "11",
                          "11",
                          "12",
                          "12"
                        ],
                        "characters": [
                          "1",
                          "2"
                        ],
                        "gameMasterFirstName": "Logan",
                        "gameMasterLastName": "Lucky",
                        "gameMasterEmail": "ll@llmail.com"
                      }
                    ]
        404:
          $ref: '#/components/responses/NotFoundError'
        default:
          $ref: '#/components/responses/UnexpectedError'

    put:
      tags:
        - Campaigns
      summary: Updates a specific campaign's name
      description: >
        Updates a the name of the specific campaign for the id provided
      operationId: updateCampaignName
      parameters:
        - name: id
          in: path
          description: ID of requested campaign
          required: true
          schema:
            type: integer
          examples:
            normal:
              summary: Normal
              value: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
            examples:
              normal:
                summary: Normal
                value: >
                  {
                    "name": "Kate's Game"
                  }
      responses:
        204:
          description: Resource updated
        404:
          description: An error occurred
      
    delete:
      tags:
        - Campaigns
      summary: Deletes a campaign
      description: >
        Deletes a campaign for the id given
      operationId: deleteCampaign
      parameters:
        - name: id
          in: path
          description: ID of campaign
          required: true
          schema:
            type: integer
          examples:
            normal:
              summary: Normal
              value: 1
      responses:
        204:
          description: Resource deleted
        404:
          $ref: '#/components/responses/NotFoundError'
        default:
          $ref: '#/components/responses/UnexpectedError'

# ------------------------------------- #
# Sessions
# ------------------------------------- #
...
components:
  responses:
    BadRequestError:
      description: The server could not understand the request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            normal:
              summary: Normal
              value: >
                {
                  "code": "BadRequest",
                  "message": "Very detailed error info"
                }
    NotFoundError:
      description: The requested resource could not be found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            normal:
              summary: Normal
              value: >
                {
                  "code": "NotFound",
                  "message": "Very detailed error info"
                }
    UnexpectedError:
      description: An unexpected error occurred
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            normal:
              summary: Normal
              value: >
                {
                  "code": "Internal",
                  "message": "Very detailed error info"
                }
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string